util: remove test-only metrics from header
It's easy for these metrics to accidentally be included in any
metrics-collection that is based off of the Kudu binaries, so I moved
them into the test file in which they're used.

Reviewed-on: http://gerrit.cloudera.org:8080/14919
Tested-by: Kudu Jenkins
